Move Wayland files into `wayland` module
