+++ src/wayland/interface.rs
@@ -1,4 +1,4 @@
-use crate::{buffer::Buffer, keyboard::KeyboardMapping, state::LockState};
+use crate::wayland::{buffer::Buffer, keyboard::KeyboardMapping, state::LockState};
 
 use wayland_client::{
     Connection, EventQueue,
